Refinement and Error Correction in the PatternAnnotator.py
I encountered an issue, that my last sheet_object was not appended to the final list of sheets.
Moreover, i recognized when searching for consecutive positions with distance == 2, when the next evaluated residue did not fullfill the condition, i started a new sheet group. But the following residues could fullfill this condition, so i adapted the code and search in all already assigned sheets, if the current residue may be appended to one of these groups, if distance == 2